design:#41-Nav bar design
