+++ src/App.jsx
@@ -1,14 +1,9 @@
-import Header from './components/Header';
 import Home from './pages/Home';
-import Footer from './components/Footer';
-import Hero from './components/Hero';
 
 function App() {
   return (
     <div>
-      <Header />
       <Home />
-      <Footer />
     </div>
   );
 }
