+++ src/components/Home/Header.jsx
@@ -1,0 +1,10 @@
+import Nav from './Nav';
+
+const Header = () => {
+  return (
+    <div>
+      <Nav />
+    </div>
+  );
+};
+export default Header;
