+++ src/components/Home/Notification.jsx
@@ -1,0 +1,32 @@
+const Notification = () => {
+  return <div>Notification</div>;
+};
+export default Notification;
+
+// import { useRecoilState } from 'recoil';
+// import { useState } from 'react';
+// import { notificationState } from '../../recoil/RecoilNotification';
+// import {
+//   Bell,
+//   NotificationContainer,
+//   Sidebar,
+//   Close,
+//   Arrow,
+// } from '../styles/NotificationCSS';
+
+// const Notification = () => {
+//   const [isOpen, setIsOpen] = useState(false);
+//   const [notification, setNotification] = useRecoilState(notificationState);
+
+//   const toggleSidebar = () => {
+//     setIsOpen(!isOpen);
+//   };
+
+//   return (
+//     <NotificationContainer>
+//       <Bell onClick={toggleSidebar} />
+//       {isOpen && <Sidebar></Sidebar>}
+//     </NotificationContainer>
+//   );
+// };
+// export default Notification;
