+++ src/try/styles(styled-components)/NotificationCSS.jsx
