+++ src/try/styles(styled-components)/SignUpMobileCSS.jsx
